fix(aws-cdk): CLI doesn't work on Node 16
During the migration, we updated `cdk-from-cfn` to the latest version.
Between version `0.162.1` and `0.189.0` a change to that package made it
incompatible with Node 16 (requires Node 18+).

Even though both Node 16 and Node 18 are EOL right now and everyone
should be on Node 20+, just to be safe we'll roll back to an old
version.

This does mean that we have an old version of the spec used by `cdk
migrate`.