chore: trigger transitive releases when packages update
For packages that include other packages, like the CLI does, it should
also include changes to bundled packages when looking at whether a
release should be triggered.

This includes `cloud-assembly-schema` and `cloudformation-diff` into
the release trigger of the CLI and all packages that include the CLI.

You'd think this could perhaps be done more declaratively, but I can't
rely on storing information on the `TypeScriptWorkSpace` project (like
"considered dependencies") because we need to pass this information
to its constructor. So I'm just copy/pasting long arrays.