feat(cdk-assets): release official GA
Remove the prerelease tag from `cdk-assets`. This tool is ready
to be used for real.

We are still releasing under the `v3-latest` tag, so CDK Pipelines users
won't automatically switch over to it just yet, but this is sending
the right signal about the tool's stability and fitness for purpose.