feat: assets can have a display name
Add a new field to the asset schema: `displayName`.

If supplied, it will be used instead of the asset ID for display
purposes in `cdk-assets` and in the CLI.

This adds support for display names to the contract and to the CLI.  A
future PR in `aws-cdk-lib` will make it possible to configure display
names on Asset objects, and add the same support to CDK Pipelines.