feat(toolkit): watch operation can be stopped
The `toolkit.watch()` operation used to start a filesystem watcher in
the background and immediately return. The caller would continue running
to its end, while the filesystem watcher would keep the node process
alive.

Instead, now return an `IWatcher` object that can be disposed
to stop the watching, and can be used to wait for the watcher to stop.