chore: retry .NET CLI mutex errors
An attempt to get rid of mutex errors in the CLI tests.

On CI we sometimes see:

    System.IO.IOException: The system cannot open the device or file specified. : 'NuGet-Migrations'

This error can be caused by lack of permissions on a temporary directory,
but in our case it's intermittent so my guess is it is caused by multiple
invocations of the .NET CLI running in parallel, and trampling on each
other creating a Mutex.

There is no fix, and it is annoyingly breaking our CI regularly. Retry a
couple of times to increase reliability.

If we still see this error after this change, maybe it is about file
permissions after all.