chore: use `>=` for dependencies on `cloud-assembly-schema`
By default we would use `^41.0.0` to depend on this package, but it is
strictly backwards compatible. The major version number of this package
currently tracks the revision of the cloud assembly protocol, and does
not use semver major versions to indicate breaking changes (breaking
changes are not allowed).

We should use `>=` to indicate that we expect a minimum version of
the schema, but we don't want to imply a maximum version.

Relates to https://github.com/aws/aws-cdk-cli/issues/345.