fix: Symbol.asyncDispose is not defined on the current JavaScript platform
We have a polyfill file that should be imported before
`Symbol.asyncDispose` is used anywhere. The polyfill file is imported in
the entry point of `@aws-cdk/toolkit-lib`, but we can't guarantee that
all imports go through the entry point.

Instead, import the polyfill in every file where `Symbol.asyncDispose`
is used.