chore(cli-integ): node-gyp fails on newer python versions
The `cli-integ` tests depend on the `node-pty` package for testing of interactive terminals. This package needs to be build at install time using `node-gyp`. Because previously `cli-integ` tests dependet on an old version of `npm`, we got an old version of `node-gyp` which is incomaptible with recent python versions.

To fix this, we upgrade the dependency to a more recent version of `npm`.