chore: shrinkwrap dependencies of @aws-cdk-testing/cli-integ
This is necessary for testing runs that run directly after a release
of a web of packages that depend on each other using exact versions
(notably, SDKv3). The problem occurs on all package managers, but
is most notable when using CodeArtifact due to slow(er) replication
latency, with an SLA of up to 30 minutes.

If we run immediately after some new package versions have gotten
released that depend on each other with precise versions , that some
packages are available under the new versions, and some packages are
not. This leads to an installation error in the tests.

By using a shrinkwrap file, at least NPM users will be saved from
this, because this will prevent the install from trying to use
the just-released package. Instead it will use the tested
combination from the shrinkwrap file.