feat(cli): metadata can live in a separate file
When CDK apps grow extremely large (think 10-100 stacks, 1000-10000
constructs), all metadata together begins to exceed 512MB, the maximum
string size in NodeJS. People usually deal with this by disabling
metadata, but they shouldn't have to.

In addition, even for manifests that don't exceed 512MB the extremely
large size of the single JSON object slows down its writing and reading
every time, even if the metadata doesn't need to be accessed.

An effective solution is to write the metadata of an artifact to a
separate file. This PR introduces the ability for that into the Cloud
Assembly schema, and updates the CLI to read from both sources if
available.

Relates to https://github.com/aws/aws-cdk/pull/34480.