fix(toolkit-lib)!: don't mutate environment when using `fromAssemblyBuilder`
The Toolkit communicates with the CDK app passing environment variables
(containing context, and others).

If the CDK app is executed as a subprocess it is trivial to set those
environment variables when executing the subprocess; but for an
in-memory synthesis as allowed by `fromAssemblyBuilder` mutating
`process.env` is not the correct approach: it is clobbering global
state, and multiple async assembly builders might fight with
each other over the single shared global object.

In `fromAssemblyBuilder()`, we now pass an additional `env` prop
containing the environment variables we didn't apply. There were
already parametes for `outdir` and `context`; those *must* now be
passed to an `App` instantiated inside the builder (whereas previously
the `App` would pick up on the globally mutated `process.env`
changes).

The rest of this PR is reorganizing the code to make the code
that touches the environment return values instead of directly
applying them.