refactor(cli): create ProxyAgent in the CLI in favor of removal from toolkit-lib
Updates the CdkToolkit and Notices APIs to take a Node https.Agent instead of options. This gives more flexibility to integrators while retaining proxy support.

The CLI changed need to be split out from the toolkit-lib changes, so the latter can be safely marked as breaking without touching CLI files.