feat(toolkit-lib)!: parameterize context storage
Previously, context was always being written to `cdk.context.json` in
some directory, even for `fromAssemblyBuilder()` or
`fromAssemblyDirectory()` which should probably both not write context
anywhere.

In this PR, we replace the context parameter to the toolkit with the
concept of a `contextStore`, which has two operations: `read()` and
`update()`.

There are simple file and in-memory implementations for simple cases, as
well as the more complex CDK app case which is implemented as
`CdkAppMultiContext`. That last one is the default for `fromCdkApp()`.

The complex one is implemented in terms of the prior
`Context/ContextBag/Settings` classes, which are probably overly
complex for the job this class is doing but keeping the old classes
is an easy way to make sure that the behavior stays compatible.