fix(cloud-assembly-schema): add a valid example for CcApiContextQuery
This type is being exposed as-is in the CDK library, by copying it into
the public API surface. It therefore gets processed by jsii, and
examples get generated for it.

Because we don't have an example for this type and jsii cannot express
string literal types, the automatically generated example is based
off the type `string` and looks like this:

```ts
{
  expectedMatchCount: 'expectedMatchCount',
}
```

This example does not compile because the generated value does not
conform to the actual declared type.

To solve this, we add a valid example here into the base code.