feat!: add `MappingSource` to the refactor API
Introduce a new type, `MappingSource`, that allows the user to tell how they want the toolkit to get the mappings: computing it, using an explicitly provided one, or using the reverse of an explicitly provided one.

This replaces three previous separate properties: `exclude`, `mappings` and `revert` with a single `mappingSource`.

This is a breaking change.