fix(integ-testing): proxy+notices test sometimes fails
The test that checks that notices are fetched through the proxy
sometimes fails; it is correctly cleaning the `notices.json` file from
a temporary directory before running the command that should fetch
notices, but the temporary directory is wrong.

When running the command it is setting `CDK_HOME` to a directory, and
`CDK_HOME` will be used to determine the location of the notices cache;
but when removing the file it is removing it from the user's home
directory, as if `CDK_HOME` was never set.

What's not clear to me is why this wouldn't *always* fail... but
I'm pretty sure the new behavior is correct.