fix(integ-runner): integ-runner produces snapshot that doesn't validate
The integ-runner uses `synthFast` to produce the assembly that will
be used to compare the snapshot to, to determine if there is "a change".

However, when a snapshot has changed and a test is run, there is a fork:

- In dry-run mode: `synthFast` will be used to produce the updated
  snapshot.
- In real-mode, the result of `this.deploy()` will be the updated
  snapshot (but also: the next time it will be compared to the result
  of `synthFast`).

We are running into a situation where the results of `synthFast` and
`this.deploy` are different, and snapshot validation fails.

In this change, we always generate the snapshot using `synthFast`, and
ignore the template produced by `this.deploy()`. This makes sure
that the templates will at least compile equal.

I'm not sure why this ever worked this way, because the template
that is deployed can be specialized to an account and region, but
the snapshot can never be, so we need to investigate a little more.