fix(schema): `unconfiguredBehavesLike` contains info for multiple versions
The way it's currently specified, the `unconfiguredBehavesLike` field
that's in the Cloud Assembly Schema requires the CLI to know the current
major version of the CDK Library to properly evaluate:

It looks like `unconfiguredBehavesLike: { v1: ..., v2: ... }` subfields,
and that requires the CLI to know whether to look in the `v1` or `v2`
subfield (and in the future `v3`, `v4`, etc...).

That is unnecessary: since the CDK Library knows its own current version
so it could just have communicated `unconfiguredBehavesLike: true` to
accurately reflect its current behavior, without the consumer needing
to know anything about the version.

The CLI code right now is only reading the `v2` field, but that will
become inaccurate if we ever release CDKv3.

In this PR, we're saying that the `v2` subfield is the "official"
location of this value going forward. `v1` should be ignored if
present (but is still included to not trip up any JSON Schema
validation).

This leads to the least amount of interference with the existing ecosystem
(no forced upgrade by CDK users and no backwards compatibility path
in the CLI to potentially read 2 fields).