fix(bootstrap): disallow AssumeRole with ExternalId by default
By default, CDK Bootstrap roles are not designed to be deputized.

(Deputized means that you give an external entity access to assume roles
on your behalf. They will supply an ExternalId to avoid [Confused Deputy
attacks](https://docs.aws.amazon.com/IAM/latest/UserGuide/confused-deputy.html))

If a deputy system (i.e., a system that assumes IAM Roles on behalf of
its tenants) is using CDK and its policies are not configured carefully,
it can be tricked into assuming its own CDK roles.

Because CDK Roles are not intended to be used in this way, we are adding
a default security control that will make this misconfiguration less
likely: AssumeRole calls with ExternalIds will be denied by default.

What if I do want to use ExternalIds?
-------------------------------------

If you are currently passing `ExternalId`s in an `AssumeRole` call to
CDK bootstrap roles *inside your own trusted organization* (expecting the
ExternalId to be present but ignored), this protection can be disabled
by calling:

```
cdk bootstrap --no-deny-external-id
```

If you want to give permissions for other organizations to assume your
CDK bootstrap roles in a deputized way, customize the bootstrap template
and add a proper `ExternalId` condition.