fix(integ-runner): region is not passed to CDK App
When using the `toolkit-lib` deployment engine, the selected region
was not being passed to the CDK app.

In the old engine, we could set the `AWS_REGION` environment variable
and the CDK CLI would respect that, but the toolkit lib was passing that
environment variable directly to the CDK app which was not respecting
that. Fix that by threading the region through as an explicit parameter
and adding an override parameter to the `AwsCliCompatible` base
credentials: we don't necessarily want to read this only from the INI
files.

Additionally: add a validation to make sure that the CDK app respects
the region we are requesting. CDK integ tests should either not
pass an environment, or read `$CDK_DEFAULT_REGION` to determine the
current region. It was too easy to just override the region with a
hardcoded one, but that breaks the integ runner's attempts to isolate
tests across regions. Guard against that by validating.