fix(integ-runner): disable update workflow by default
I've lost 2 hours
debugging why my deployment was failing trying to deploy to nonexistent
and non-requested accounts and regions, and the
answer was: the checked-in snapshot was nonsensical, and the update
workflow will first deploy the checked-in snapshot and only then
deploy the current snapshot.

Good idea, bad execution: if existing
snapshots are not region-agnostic they are guaranteed to fail. We
should probably resynthesize them from the previous git commit instead.

Whatever it is, what we do right now doesn't work so I'm disabling it.